refactor/move helper fns to separate file
